Changed to int writabel from long

+++ src/main/java/com/innovisto/hadoop/mrdemo/email/EmailProcessRunner.java
@@ -4,7 +4,7 @@ import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.conf.Configured;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
-import org.apache.hadoop.io.LongWritable;
+import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Job;
 import org.apache.hadoop.mapreduce.lib.input.FileInputFormat;
@@ -40,7 +40,7 @@ public class EmailProcessRunner extends Configured implements Tool {
 
         // Specify key / value
         job.setOutputKeyClass(Text.class);
-        job.setOutputValueClass(LongWritable.class);
+        job.setOutputValueClass(IntWritable.class);
 
         // Input
         FileInputFormat.addInputPath(job, new Path(args[0]));
